Reorganize RPC structure
