+++ cmd/test/backend_kv_test.cpp
@@ -37,9 +37,9 @@
 #include <silkworm/common/log.hpp>
 #include <silkworm/common/util.hpp>
 #include <silkworm/db/tables.hpp>
-#include <silkworm/rpc/completion_tag.hpp>
-#include <silkworm/rpc/conversion.hpp>
-#include <silkworm/rpc/util.hpp>
+#include <silkworm/rpc/common/completion_tag.hpp>
+#include <silkworm/rpc/common/conversion.hpp>
+#include <silkworm/rpc/common/util.hpp>
 
 using namespace std::literals;
