+++ node/silkworm/backend/rpc/backend_calls.cpp
@@ -22,8 +22,8 @@
 #include <types/types.pb.h>
 
 #include <silkworm/common/log.hpp>
-#include <silkworm/rpc/conversion.hpp>
-#include <silkworm/rpc/util.hpp>
+#include <silkworm/rpc/common/conversion.hpp>
+#include <silkworm/rpc/common/util.hpp>
 
 namespace silkworm::rpc {
