+++ node/silkworm/rpc/common/util.hpp
